fix: the slice is typed and divided into folders part 2
