+++ src/redux/slices/hooks.ts
@@ -1,0 +1,5 @@
+import {useDispatch, useSelector, TypedUseSelectorHook} from 'react-redux'
+import {AppDispatch, RootState} from '../store';
+
+export const useAppDispatch = () => useDispatch<AppDispatch>()
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
