+++ src/redux/slices/mainSlice/reducers/mainReducer.ts
@@ -1,3 +1,13 @@
-export const setItemsReducer = (state: any, action: any) => {
-  state.data = action.payload;
-};
+import { PayloadAction } from "@reduxjs/toolkit";
+import { Status, ToDo } from "../types/mainTypes";
+
+export const addToDo = (state: ToDo[], action: PayloadAction<string>) => {
+  const newToDo: ToDo = {
+    id: new Date().toString(),
+    title: action.payload,
+    error: null,
+    status: Status.SUCCESS
+  }
+
+  return [newToDo, ...state]
+}
