--- src/redux/slices/mainSlice/selectors/mainSelector.ts
@@ -1,3 +1,0 @@
-import { RootState } from "../../../store";
-
-export const selectToDoData = (state: RootState) => state.data;
